Refactor deprecation check to shared crdmanagement package
For future re-use
